ci: clippy fixes [ci skip]
